Updated CategoryForm correcting url oversight and text fields, updated CategoryOverviewData balance val to double and will likely do elsewhere as testing continues

+++ shared/src/commonMain/kotlin/com/budgetbud/kmp/models/CategoryOverviewData.kt
@@ -6,5 +6,5 @@ import kotlinx.serialization.Serializable
 data class CategoryOverviewData(
     val id: Int,
     val name: String,
-    val balance: Long
+    val balance: Double
 )
